feat: Consolidate grid shape extensions and align Standard with Native version
- Moved shape factory methods and transformation logic to generated extensions
- Added CheckShapeCells method to template with zero-allocation delegate pattern
- Added WritableGridShapeExtension for mutable operations (SetCellValue, FillAll, FillRect, FillShape, Clear, PlaceItem, RemoveItem)
- Updated Standard version to match Native API structure:
  - Added GetIndex, GetCell, SetCell methods to GridShape and GridShape.ReadOnly
  - Added math wrapper class for Standard version (min, max, tzcnt, lzcnt)
  - Added Allocator enum stub for API compatibility
  - Fixed indexer patterns to use GetIndex for better compatibility
  - Updated ImmutableGridShape to use GetCellValue extension methods
  - Added freeValue parameter to IsTrimmed, CanPlaceItem, FindFirstFit methods
- Removed separate Shapes.cs files (now generated in extensions)
- Removed GridBoardExtension.cs and ShapesValueGridExtension.cs (consolidated)
- Updated GridBoard implementations to use new extension methods
- Fixed ValueGridShape indexers to use extension methods
