--- Packages/com.fullmetalbagel.dope-grid/Runtime/Native/EditorExtension.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 70adb77c725b4195a9e3b76d69f34e32
-timeCreated: 1758436457
