--- Packages/com.fullmetalbagel.dope-grid/Runtime/Native/Shapes.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 28ddd48c79ab4bc2a0f5abab21f3a867
-timeCreated: 1757667638
