--- Packages/com.fullmetalbagel.dope-grid/Runtime/Standard/GridBoardExtension.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 79646617fe15f41ada38051dc978e3de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
