--- Packages/com.fullmetalbagel.dope-grid/Runtime/Standard/Shapes.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f017c0373f5644e7f80da79019c85fc4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
