--- Packages/com.fullmetalbagel.dope-grid/Runtime/Standard/ShapesValueGridExtension.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47e6215e6ee4848039efb1720f7ee39f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
