+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Standard/ValueGridShape.cs
@@ -32,22 +32,16 @@ public struct ValueGridShape<T> : IEquatable<ValueGridShape<T>>, IDisposable whe
         Array.Fill(_values, defaultValue, 0, width * height);
     }
 
-    public T GetValue(GridPosition pos) => GetValue(pos.X, pos.Y);
-    public T GetValue(int x, int y) => Values[GetIndex(x, y)];
-
-    public void SetValue(GridPosition pos, T value) => SetValue(pos.X, pos.Y, value);
-    public void SetValue(int x, int y, T value) => Values[GetIndex(x, y)] = value;
-
     public T this[int x, int y]
     {
-        get => GetValue(x, y);
-        set => SetValue(x, y, value);
+        get => this.GetCellValue(x, y);
+        set => this.SetCellValue(x, y, value);
     }
 
     public T this[GridPosition pos]
     {
-        get => GetValue(pos);
-        set => SetValue(pos, value);
+        get => this.GetCellValue(pos);
+        set => this.SetCellValue(pos, value);
     }
 
     public T this[int index]
@@ -56,39 +50,10 @@ public struct ValueGridShape<T> : IEquatable<ValueGridShape<T>>, IDisposable whe
         set => Values[index] = value;
     }
 
-    public void Fill(T value)
+    public void FillAll(T value)
     {
         Values.Fill(value);
     }
-
-    public void FillRect(int x, int y, int width, int height, T value)
-    {
-        for (int dy = 0; dy < height; dy++)
-        {
-            for (int dx = 0; dx < width; dx++)
-            {
-                var px = x + dx;
-                var py = y + dy;
-                if (px >= 0 && px < Width && py >= 0 && py < Height)
-                {
-                    SetValue(px, py, value);
-                }
-            }
-        }
-    }
-
-    public void FillRect(GridPosition pos, (int width, int height) size, T value)
-    {
-        FillRect(pos.X, pos.Y, size.width, size.height, value);
-    }
-
-    public void Clear()
-    {
-        Fill(default!);
-    }
-
-    public bool Contains(GridPosition pos) => Contains(pos.X, pos.Y);
-    public bool Contains(int x, int y) => x >= 0 && x < Width && y >= 0 && y < Height;
 
     public ValueGridShape<T> Clone() => AsReadOnly().Clone();
     public void CopyTo(ValueGridShape<T> other) => AsReadOnly().CopyTo(other);
@@ -105,7 +70,7 @@ public struct ValueGridShape<T> : IEquatable<ValueGridShape<T>>, IDisposable whe
         {
             for (int x = 0; x < Width; x++)
             {
-                SetValue(x, y, shape.GetCell(x, y) ? trueValue : falseValue);
+                this[x, y] = shape[x, y] ? trueValue : falseValue;
             }
         }
     }
@@ -153,14 +118,8 @@ public struct ValueGridShape<T> : IEquatable<ValueGridShape<T>>, IDisposable whe
             _values = values;
         }
 
-        public int GetIndex(GridPosition pos) => GetIndex(pos.X, pos.Y);
-        public int GetIndex(int x, int y) => y * Width + x;
-
-        public T GetValue(GridPosition pos) => GetValue(pos.X, pos.Y);
-        public T GetValue(int x, int y) => _values[GetIndex(x, y)];
-
-        public T this[int x, int y] => GetValue(x, y);
-        public T this[GridPosition pos] => GetValue(pos);
+        public T this[int x, int y] => this.GetCellValue(x, y);
+        public T this[GridPosition pos] => this.GetCellValue(pos);
         public T this[int index] => _values[index];
 
         public ValueGridShape<T> Clone()
@@ -194,7 +153,7 @@ public struct ValueGridShape<T> : IEquatable<ValueGridShape<T>>, IDisposable whe
             {
                 for (int x = 0; x < Width; x++)
                 {
-                    shape.SetCell(x, y, predicate(GetValue(x, y), data));
+                    shape[x, y] = predicate(this[x, y], data);
                 }
             }
             return shape;
